ai-service/requirements.txt
flask
google-generativeai
pandas
scikit-learn
python-dotenv
flask-cors
PyPDF2
